# backend/orders-api/src/main/java/the/chak/ecommerce/orders/boundary/dto/ProductVO.java
package the.chak.ecommerce.orders.boundary.dto;

import lombok.Data;

@Data
public class ProductVO {
    private String productID;
    private String title;
    private Integer qty;
    private Double price;
    private Double percentageOff;
}