# backend/orders-service/src/main/java/the/chak/ecommerce/orders/boundary/dto/CartResponse.java
package the.chak.ecommerce.orders.boundary.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

import java.time.Instant;
import java.util.List;

@Data
@AllArgsConstructor
public class CartResponse {
    private String userId;
    private List<CartItemResponse> items;
    private Instant updatedAt;
}